README and testing working
