--- hardhat_test/.env.example
@@ -1,2 +1,0 @@
-# .env
-PRIVATE_KEY =  " YOUR_KEY_HERE"
