--- hardhat_test/README.md
@@ -1,22 +1,0 @@
-# Sample Hardhat Project
-
-This project demonstrates a basic Hardhat use case. It comes with a sample contract, a test for that contract, and a script that deploys that contract.
-
-Try running some of the following tasks:
-
-```shell
-npx hardhat help
-npx hardhat test
-GAS_REPORT=true npx hardhat test
-npx hardhat node
-
-npx hardhat run scripts/deploy.js
-npx hardhat run scripts/deploy.js --network goerli 
-
-if you only want to call functions and not redeploy
-npx hardhat run scripts/trigger_proofs.js --network goerli  
-
-Integration testing for API: 
-npx hardhat test --network goerli
-
-```
