--- hardhat_test/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "hardhat_test",
-  "version": "1.0.0",
-  "description": "This project demonstrates a basic Hardhat use case. It comes with a sample contract, a test for that contract, and a script that deploys that contract.",
-  "main": "hardhat.config.js",
-  "directories": {
-    "lib": "lib"
-  },
-  "dependencies": {
-    "@chainlink/contracts": "^0.4.1",
-    "@nomicfoundation/hardhat-toolbox": "^1.0.2",
-    "axios": "^0.27.2",
-    "balanced-match": "^1.0.2",
-    "brace-expansion": "^1.1.11",
-    "chai": "^4.3.6",
-    "coffeescript": "^1.12.7",
-    "concat-map": "^0.0.1",
-    "dotenv": "^16.0.1",
-    "fs.realpath": "^1.0.0",
-    "gaze": "^1.1.3",
-    "globule": "^1.3.4",
-    "growl": "^1.10.0",
-    "hardhat": "^2.10.2",
-    "inflight": "^1.0.6",
-    "inherits": "^2.0.4",
-    "jasmine-growl-reporter": "^1.0.3",
-    "jasmine-node": "^1.16.0",
-    "jasmine-reporters": "^1.0.2",
-    "lodash": "^4.17.21",
-    "mkdirp": "^0.3.5",
-    "mocha": "^10.0.0",
-    "once": "^1.4.0",
-    "path-is-absolute": "^1.0.1",
-    "request": "^2.88.2",
-    "requirejs": "^2.3.6",
-    "underscore": "^1.13.4",
-    "walkdir": "^0.0.12",
-    "wrappy": "^1.0.2",
-    "xhr2": "^0.2.1"
-  },
-  "devDependencies": {
-    "jasmine": "^4.3.0"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
-  "author": "",
-  "license": "ISC"
-}
