--- validate.json
@@ -1,7 +1,0 @@
-{
-    "id": "12345",
-    "data":
-    {
-        "deal_id": "1"
-    }
-}
